Add back removed other repo owners in GH docker actions
In 008a2f68694812b7, support was added for using Docker related Github
actions in repositories of other owners. This was pretty helpful as it
allowed running modified forks off of main in a nightly fashion before
patches were pushed upstream. This was 45c83f73aca11333c4c47, add it
back